// client/src/Gen/models/MealFoodItemEntryPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nutrition Tracker API
 * API Specification for Nutrition Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface MealFoodItemEntryPostRequest
 */
export interface MealFoodItemEntryPostRequest {
    /**
     *
     * @type {number}
     * @memberof MealFoodItemEntryPostRequest
     */
    amount: number;
    /**
     *
     * @type {number}
     * @memberof MealFoodItemEntryPostRequest
     */
    foodItemId: number;
}

/**
 * Check if a given object implements the MealFoodItemEntryPostRequest interface.
 */
export function instanceOfMealFoodItemEntryPostRequest(value: object): value is MealFoodItemEntryPostRequest {
    if (!("amount" in value) || value["amount"] === undefined) return false;
    if (!("foodItemId" in value) || value["foodItemId"] === undefined) return false;
    return true;
}

export function MealFoodItemEntryPostRequestFromJSON(json: any): MealFoodItemEntryPostRequest {
    return MealFoodItemEntryPostRequestFromJSONTyped(json, false);
}

export function MealFoodItemEntryPostRequestFromJSONTyped(
    json: any,
    ignoreDiscriminator: boolean,
): MealFoodItemEntryPostRequest {
    if (json == null) {
        return json;
    }
    return {
        amount: json["amount"],
        foodItemId: json["foodItemId"],
    };
}

export function MealFoodItemEntryPostRequestToJSON(json: any): MealFoodItemEntryPostRequest {
    return MealFoodItemEntryPostRequestToJSONTyped(json, false);
}

export function MealFoodItemEntryPostRequestToJSONTyped(
    value?: MealFoodItemEntryPostRequest | null,
    ignoreDiscriminator: boolean = false,
): any {
    if (value == null) {
        return value;
    }

    return {
        amount: value["amount"],
        foodItemId: value["foodItemId"],
    };
}